refactor: loginGuard 로직 변경
