+++ frontend/techpick-extension/src/components/index.ts
@@ -1,6 +1,5 @@
 export { DeleteTagDialog } from './DeleteTagDialog';
 export { DeselectTagButton } from './DeselectTagButton';
-export { LoginGuard } from './LoginGuard';
 export { ToggleThemeButton } from './ToggleThemeButton';
 export { TagPicker } from './TagPicker';
 export { SkeltonPickForm } from './SkeltonPickForm';
